requirements.txt
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
loguru==0.7.0
outcome==1.2.0
praw==7.7.0
prawcore==2.3.0
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
selenium==4.10.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.10.3
update-checker==0.18.0
urllib3==2.0.3
websocket-client==1.5.3
wsproto==1.2.0
